pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ti_kit_board_communication"
description = "This project creates a web app using the streamlit Python library to interact with the TiKit board visually."
readme = "README.md"
version = "0.4.0"
requires-python = ">=3.13.0"
dependencies = ["pyserial"]
authors = [
    { name = "Vishwesswaran Gopal", email = "vishy1290@gmail.com" },
]
maintainers = [
    { name = "Vishwesswaran Gopal", email = "vishy1290@gmail.com" },
]
keywords = ["embedded systems", "python", "MSP-EXP430FR5994", "TiKit"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Embedded Systems",
    "Topic :: Software Development :: Libraries"
]

[tool.hatch.build.targets.sdist]
packages = ["ti_kit_board_communication"]

[tool.hatch.build.targets.wheel]
packages = ["ti_kit_board_communication"]
